requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pdfplumber
python-docx